Updated RoboCOP. Added many changes: 1. Calculate dinucleotide pwm of nucleosome from input allowing other genomes. 2. Saving output as h5py. 3. Input of PWMs in MEME format 3. Standardizing inputs of other files to be in BED format. 4. Made plotting function more flexible. 5. Added tutorial vignette and a toy BAM file

+++ pkg/robocop_em.py
@@ -1,7 +1,5 @@
 import os
 import sys
-sys.path.append("/home/home3/sneha/.local/lib/python3.6/site-packages/")
-sys.path.append(os.getcwd())
 import robocop
 from robocop.utils.readWriteOps import *
 from robocop.utils.robocopExtras import *
@@ -9,20 +7,23 @@ import robocop.utils.parameterize as parameterize
 import numpy as np
 import pickle
 import re
-from configparser import SafeConfigParser
+from configparser import ConfigParser
 import robocop.utils.getReads as getReads
+import robocop.utils.readData as readData
 import pandas
 import gc
-from multiprocessing import Pool, Manager
+import h5py
+import random
+random.seed(9)
 
 # create posterior table for each segment
-def createInstances(tf_prob, dbf_conc, coords, pwm, cshared, tmpDir, nucleotide_sequence, mnaseParams, tech, pool):
+def createInstances(tf_prob, dbf_conc, coords, pwm, cshared, tmpDir, info_file, fasta_file, nucleosome_file, nucleotide_sequence, mnaseParams, tech):
     segments = len(coords)
     dshared = {} 
     dshared["robocopC"] = cshared 
-    robocop.createSharedDictionary(dshared, tf_prob, dbf_conc['background'], dbf_conc['nucleosome'], pwm, tmpDir, nucleotide_sequence)
+    robocop.createSharedDictionary(dshared, fasta_file, nucleosome_file, tf_prob, dbf_conc['background'], dbf_conc['nucleosome'], pwm, tmpDir, info_file, nucleotide_sequence)
     for t in range(segments):
-        createInstance((t, dshared))
+        createInstance((t, dshared, coords.iloc[t]['chr'], coords.iloc[t]['start'], coords.iloc[t]['end']))
     if mnaseParams != None:
         for s in range(segments):
             updateMNaseEMMatNB((s, dshared, mnaseParams, tech))
@@ -31,8 +32,9 @@ def createInstances(tf_prob, dbf_conc, coords, pwm, cshared, tmpDir, nucleotide_
     gc.collect()
     return dshared
     
-def runROBOCOP_EM(coordFile, config, outDir, tmpDir, pool, mnaseFile, dnaseFiles = ""):
+def runROBOCOP_EM(coordFile, config, outDir, tmpDir, info_file_name, mnaseFile, dnaseFiles = ""):
 
+    info_file = h5py.File(info_file_name, mode = 'w') 
     fragRangeLong = tuple([int(x) for x in re.findall(r'\d+', config.get("main", "fragRangeLong"))])
     fragRangeShort = tuple([int(x) for x in re.findall(r'\d+', config.get("main", "fragRangeShort"))])
     fragRange = (fragRangeLong, fragRangeShort)
@@ -43,13 +45,25 @@ def runROBOCOP_EM(coordFile, config, outDir, tmpDir, pool, mnaseFile, dnaseFiles
     # chromosome segments in pandas data frame
     coords = pandas.read_csv(coordFile, sep = "\t")
 
+    # select suubset for training 
+    if len(coords) > 500: 
+        idx = list(range(len(coords)))
+        random.shuffle(idx)
+        idx = idx[:500]
+        coords = coords.iloc[idx]
+        coords = coords.reset_index()
+        
     # dbf weights initlaized using KD values
-    dbf_conc, pwm = parameterize.getDBFconc(nucFile, config.get("main", "pwmFile"))
+    dbf_conc, pwm = parameterize.getDBFconc(nucFile, config.get("main", "pwmFile"), outDir)
 
     # read nucleotide sequence and return 1 if successful
-    nucleotide_sequence = getReads.getNucSequence(nucFile, tmpDir, coords)
+    nucleotide_sequence = getReads.getNucSequence(nucFile, tmpDir, info_file, coords)
     # read MNase-seq midpoint counts for long and short fragments
-    mnase_data_long, mnase_data_short = getReads.getMNase(mnaseFiles, tmpDir, coords, fragRange, tech = tech)
+
+    # readData.get2DValues(mnaseFiles, config.get('main', 'chrSizesFile'), (0, 200), tmpDir)
+    # mnase_data_long, mnase_data_short = getReads.getMNaseSmoothed(tmpDir, coords, fragRange, tech = tech)
+
+    mnase_data_long, mnase_data_short = getReads.getMNase(mnaseFiles, tmpDir, info_file, coords, fragRange, tech = tech)
     
     # make t copies of each tf_prob for every timepoint -- only 1 timepoint
     timepoints = len(coords)
@@ -63,28 +77,27 @@ def runROBOCOP_EM(coordFile, config, outDir, tmpDir, pool, mnaseFile, dnaseFiles
         if i != 'nucleosome' and i != 'background':
             tf_prob[i] = dbf_conc[i] 
             if threshold < tf_prob[i] and i != 'unknown':
-                threshold = tf_prob[i][0]
-            if i != 'unknown': thresholds.append(tf_prob[i][0])
-            tName.append((i, tf_prob[i][0]))
+                threshold = tf_prob[i]
+            if i != 'unknown': thresholds.append(tf_prob[i])
+            tName.append((i, tf_prob[i]))
     threshold = np.mean(thresholds) + 2*np.std(thresholds)
     # get MNase-seq count parameters
     if mnaseFiles: 
-        mnaseParams = parameterize.getParamsMNase(mnaseFiles, config.get("main", "nucleosomeFile"), config.get("main", "tfFile"), fragRange, tech)
+        # with open(outDir + "/negParamsMNase.pkl", 'rb') as readFile:
+        #     mnaseParams = pickle.load(readFile, encoding = 'latin1')
+        mnaseParams = parameterize.getParamsMNase(mnaseFiles, config.get("main", "nucleosomeFile"), config.get("main", "tfFile"), fragRange, tmpDir, tech)
     else:
         mnaseParams = None
     
     # create shared dictionary for all segments and build HMM transition matrix
-    dshared = createInstances(tf_prob, dbf_conc, coords, pwm, cshared, tmpDir, nucleotide_sequence, mnaseParams, tech, pool)
+    dshared = createInstances(tf_prob, dbf_conc, coords, pwm, cshared, tmpDir, info_file, config.get("main", "nucFile"), config.get("main", "nucleosomeFile"), nucleotide_sequence, mnaseParams, tech)
 
     fLike = open(outDir + '/likelihood.txt', 'w')
-    likelihood = getLogLikelihood(segments, dshared['tmpDir'])
+    likelihood = getLogLikelihood(segments, dshared)
     fLike.write(str(likelihood) + '\n')
     fLike.close()
     iterations = 10
     countMNase = 0
-    lfMNaseLong = 1
-    lfMNaseShort = 1
-    lfMNaseScale = 1
 
     print("Writing MNase params")
     if mnaseFiles != "":
@@ -96,8 +109,9 @@ def runROBOCOP_EM(coordFile, config, outDir, tmpDir, pool, mnaseFile, dnaseFiles
         #create new dshared
         dsharedNew = {}
         for k in list(dshared.keys()):
-            dsharedNew[k] = dshared[k]
-        print("Writing to HMMconfig")
+            if k == 'info_file': dsharedNew['info_file_name'] = info_file_name
+            else: dsharedNew[k] = dshared[k]
+        # print("Writing to HMMconfig")
         with open(outDir + "/HMMconfig" + str(i) + ".pkl", 'wb') as writeFile:
             pickle.dump(dsharedNew, writeFile, pickle.HIGHEST_PROTOCOL)
 
@@ -112,58 +126,30 @@ def runROBOCOP_EM(coordFile, config, outDir, tmpDir, pool, mnaseFile, dnaseFiles
         for t in range(segments):
             setValuesPosterior((t, dshared, tf_prob, background_prob, nucleosome_prob, tmpDir))
 
-    likelihood = getLogLikelihood(segments, dshared['tmpDir'])
-    fLike = open(outDir + '/likelihood.txt', 'a')
-    fLike.write(str(likelihood) + '\n')
-    fLike.close()
+        likelihood = getLogLikelihood(segments, dshared)
+        fLike = open(outDir + '/likelihood.txt', 'a')
+        fLike.write(str(likelihood) + '\n')
+        fLike.close()
     
     #create new dshared
     dsharedNew = {}
     for k in list(dshared.keys()):
-        dsharedNew[k] = dshared[k]
+        if k == 'info_file': dsharedNew['info_file_name'] = info_file_name
+        else: dsharedNew[k] = dshared[k]
     print("Writing to HMMconfig")
     with open(outDir + "/HMMconfig.pkl", 'wb') as writeFile:
         pickle.dump(dsharedNew, writeFile, pickle.HIGHEST_PROTOCOL)
 
     # remove tmpDir
-    os.system ("rm -rf " + outDir + "tmpDir")
+    if len(coords) <= 500: os.system ("rm -rf " + outDir + "tmpDir")
 
-if __name__ == '__main__':
-        if len(sys.argv) != 4:
-                print("Usage: python robocop_em.py <coordinate file> <config file> <output directory>")
-                exit(1)
-        coordFile = sys.argv[1]
-        configFile = sys.argv[2]
-        outDir = sys.argv[3]
     
-        print("RoboCOP: model training ...")
-        print("Coordinates: " + coordFile)
-        print("Config file: " + configFile)
-        print("Output dir: " + outDir)
+if __name__ == '__main__':
+    if len(sys.argv) != 4:
+        print("Usage: python robocop_em.py <coordinate file> <config file> <output directory>")
+        exit(1)
+    coordFile = sys.argv[1]
+    configFile = sys.argv[2]
+    outDir = sys.argv[3]
 
-        os.makedirs(outDir, exist_ok = True)
-        tmpDir = outDir + "/tmpDir/"
-        os.makedirs(tmpDir, exist_ok = True)
-        
-        # copy config file
-        os.system("cp " + configFile + " " + outDir + "/config.ini") 
-        # copy coordinates file
-        os.system("cp " + coordFile + " " + outDir + "/coords.tsv") 
-
-        configFile = outDir + "/config.ini"
-
-        fc = open(configFile, 'a')
-        try:
-            fc.write("\ntrainDir = " + outDir + "\n")
-            fc.close()
-        except configparser.DuplicateOptionError:
-            fc.close()
-
-        # read config file
-        config = SafeConfigParser()
-        config.read(configFile)
-        nProcs = int(config.get("main", "nProcs"))
-        if not config.has_option("main", "bamFile"): bamFile = None
-        else: bamFile = config.get("main", "bamFile")
-
-        runROBOCOP_EM(coordFile, config, outDir, tmpDir, 1, bamFile)
+    run_robocop_with_em(coordFile, configFile, outDir)
